refactor(users): extract after-successful-login logic to a hook
