+++ plugins/leemons-plugin-users/frontend/src/pages/public/Login/index.js
@@ -1,7 +1,6 @@
 import React, { useMemo, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import _ from 'lodash';
-import { Box, createStyles, Stack } from '@bubbles-ui/components';
 import { LoginForm } from '@users/components/LoginForm';
 import tLoader from '@multilanguage/helpers/tLoader';
 import useCommonTranslate from '@multilanguage/helpers/useCommonTranslate';
@@ -10,25 +9,9 @@ import prefixPN from '@users/helpers/prefixPN';
 import { goRecoverPage } from '@users/navigate';
 import { AuthLayout } from '@users/layout/AuthLayout';
 import { AuthContainer } from '@users/components/AuthContainer';
-import {
-  getRememberLoginRequest,
-  getUserProfilesRequest,
-  getUserProfileTokenRequest,
-  loginRequest,
-} from '@users/request';
+import { loginRequest } from '@users/request';
 import { getCookieToken, useSession } from '@users/session';
-import Cookies from 'js-cookie';
-import hooks from 'leemons-hooks';
-import { getUserCenterProfileTokenRequest, getUserCentersRequest } from '../../../request';
-
-const PageStyles = createStyles((theme) => ({
-  root: {
-    padding: theme.spacing[7],
-  },
-  content: {
-    maxWidth: 330,
-  },
-}));
+import useOnLoginSuccess from '@users/hooks/useOnLoginSuccess';
 
 const UNKNOWN_ERROR = 'unknown-error';
 
@@ -45,66 +28,12 @@ export default function Login() {
   const history = useHistory();
   const [formStatus, setFormStatus] = useState('');
   const [formError, setFormError] = useState(null);
+  const onLoginSuccess = useOnLoginSuccess();
 
   const { t: tCommon } = useCommonTranslate('forms');
 
   const [translations] = useTranslate({ keysStartsWith: prefixPN('login') });
   const t = tLoader(prefixPN('login'), translations);
-
-  const handleProfileAndCenter = async (profile, center, jwtToken) => {
-    if (profile.sysName === 'admin') {
-      const response = await getUserProfileTokenRequest(profile.id, jwtToken);
-      return { ...response.jwtToken, profile };
-    }
-    const response = await getUserCenterProfileTokenRequest(center.id, profile.id, jwtToken);
-    await hooks.fireEvent('user:change:profile', profile);
-    return response.jwtToken;
-  };
-
-  const handleNoProfileAndCenter = async (jwtToken) => {
-    const [{ centers }, { profiles }] = await Promise.all([
-      getUserCentersRequest(jwtToken),
-      getUserProfilesRequest(jwtToken),
-    ]);
-
-    if (profiles.length === 1 && profiles[0].sysName === 'admin') {
-      const response = await getUserProfileTokenRequest(profiles[0].id, jwtToken);
-      return { ...response.jwtToken, profile: profiles[0] };
-    }
-
-    if (centers.length === 1 && centers[0].profiles.length === 1 && profiles.length === 1) {
-      const response = await getUserCenterProfileTokenRequest(
-        centers[0].id,
-        centers[0].profiles[0].id,
-        jwtToken
-      );
-      await hooks.fireEvent('user:change:profile', centers[0].profiles[0]);
-
-      return response.jwtToken;
-    }
-
-    return jwtToken;
-  };
-
-  const handleLoginSuccess = async (token) => {
-    let jwtToken;
-    const { profile, center } = await getRememberLoginRequest(token);
-
-    if (profile && center) {
-      jwtToken = await handleProfileAndCenter(profile, center, token);
-    } else {
-      jwtToken = await handleNoProfileAndCenter(token);
-    }
-
-    Cookies.set('token', jwtToken);
-    hooks.fireEvent('user:cookie:session:change');
-
-    const redirectUrl = _.isString(jwtToken)
-      ? '/protected/users/select-profile'
-      : '/private/dashboard';
-
-    history.push(redirectUrl);
-  };
 
   const handleLoginError = (err) => {
     if (err.message === 'exceeded-login-attempts') {
@@ -129,9 +58,8 @@ export default function Login() {
     try {
       setFormStatus('loading');
       setFormError(null);
-      window.sessionStorage.setItem('boardMessagesModalId', null);
       const response = await loginRequest(data);
-      await handleLoginSuccess(response?.jwtToken);
+      await onLoginSuccess(response?.jwtToken);
     } catch (err) {
       console.error('error', err);
       handleLoginError(err);
@@ -168,8 +96,6 @@ export default function Login() {
     }),
     [tCommon]
   );
-
-  const { classes } = PageStyles();
 
   return (
     <AuthLayout>
